requirements.txt
gunicorn==19.9.0
psycopg2==2.8.6
Pillow==8.3.0
Django==2.2.5
django-google-maps==0.11.0
django-mptt==0.10.0
django-widget-tweaks==1.4.5
django-filter==2.2.0
django-tables2==2.1.1
djangorestframework==3.10.3
django-environ==0.7.0
django-phonenumber-field == 4.0.0
phonenumbers == 8.11.1
django-bootstrap-datepicker-plus == 3.0.5
django-taggit == 1.2.0
django-taggit-serializer == 0.1.7
djangorestframework-recursive == 0.1.2
python-slugify == 4.0.0
django-extensions==3.1.3
django-environ==0.7.0
